pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "your-project-name"
version = "0.1.0"
dependencies = [
    "accelerate",
    "ag2[openai]>=0.3.2",
    "aiohappyeyeballs",
    "aiohttp",
    "aiohttp-cors",
    "aiosignal",
    "airportsdata",
    "altair",
    "annotated-types",
    "anyio",
    "astor",
    "attrs",
    "blake3",
    "blinker",
    "cachetools",
    "certifi",
    "charset-normalizer",
    "click",
    "cloudpickle",
    "cohere",
    "coloredlogs",
    "colorful",
    "compressed-tensors",
    "dataclasses-json",
    "datasets",
    "depyf",
    "dill",
    "diskcache",
    "distlib",
    "distro",
    "einops",
    "eval_type_backport",
    "faiss-cpu",
    "fastapi",
    "fastavro",
    "filelock",
    "filetype",
    "flaml>=2.3.6",
    "flatbuffers",
    "frozenlist",
    "fsspec",
    "gdown>=5.2.0",
    "gguf",
    "gitdb",
    "GitPython",
    "google-ai-generativelanguage",
    "google-api-core",
    "google-auth",
    "googleapis-common-protos",
    "gptcache",
    "greenlet",
    "gritlm",
    "grpcio",
    "grpcio-status",
    "h11",
    "h2",
    "hipporag",
    "hpack",
    "httpcore",
    "httptools",
    "httpx",
    "httpx-sse",
    "huggingface-hub[hf-xet]",
    "humanfriendly",
    "hyperframe",
    "idna",
    "igraph",
    "importlib_metadata",
    "interegular",
    "ipykernel>=7.1.0",
    "Jinja2",
    "jiter",
    "joblib",
    "jsonpatch",
    "jsonpointer",
    "jsonschema",
    "jsonschema-specifications",
    "langchain==0.3.27",
    "langchain-cohere",
    "langchain-community",
    "langchain-core",
    "langchain-google-genai",
    "langchain-huggingface",
    "langchain-openai",
    "langchain-qdrant",
    "langchain-text-splitters",
    "langgraph",
    "langgraph-checkpoint",
    "langgraph-prebuilt",
    "langgraph-sdk",
    "langsmith",
    "lark",
    "lm-format-enforcer",
    "markdown-it-py",
    "MarkupSafe",
    "marshmallow",
    "mdurl",
    "mem0ai[async]>=0.1.115",
    "mistral_common",
    "mpmath",
    "msgpack",
    "msgspec",
    "mteb",
    "multidict",
    "multiprocess",
    "mypy_extensions",
    "narwhals",
    "nest-asyncio",
    "networkx",
    "ninja",
    "numpy",
    "nvidia-cublas-cu12",
    "nvidia-cuda-cupti-cu12",
    "nvidia-cuda-nvrtc-cu12",
    "nvidia-cuda-runtime-cu12",
    "nvidia-cudnn-cu12",
    "nvidia-cufft-cu12",
    "nvidia-curand-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-cusparse-cu12",
    "nvidia-ml-py",
    "nvidia-nccl-cu12",
    "nvidia-nvjitlink-cu12",
    "nvidia-nvtx-cu12",
    "onnxruntime==1.17.3",
    "openai",
    "opencensus",
    "opencensus-context",
    "opencv-python-headless",
    "opentelemetry-api",
    "opentelemetry-exporter-prometheus",
    "opentelemetry-proto",
    "opentelemetry-sdk",
    "opentelemetry-semantic-conventions",
    "orjson",
    "ormsgpack",
    "outlines",
    "outlines_core",
    "packaging",
    "pandas",
    "partial-json-parser",
    "pillow",
    "platformdirs",
    "polars",
    "polars-runtime-32",
    "portalocker",
    "prometheus-fastapi-instrumentator",
    "prometheus_client",
    "propcache",
    "proto-plus",
    "protobuf",
    "psutil",
    "py-cpuinfo",
    "py-spy",
    "pyarrow",
    "pyasn1",
    "pyasn1_modules",
    "pyautogen>=0.9",
    "pycountry",
    "pydantic",
    "pydantic-extra-types",
    "pydantic-settings",
    "pydantic_core",
    "pydeck",
    "Pygments",
    "python-dateutil",
    "python-dotenv",
    "python-igraph",
    "pytrec_eval-terrier",
    "pytz",
    "PyYAML",
    "pyzmq",
    "qdrant-client",
    "ray",
    "redis>=7.1.0",
    "referencing",
    "regex",
    "requests",
    "requests-toolbelt",
    "rich",
    "rpds-py",
    "rsa",
    "safetensors",
    "scikit-learn",
    "scipy",
    "sentence-transformers",
    "sentencepiece",
    "sentry-sdk",
    "setuptools",
    "six",
    "smart_open",
    "smmap",
    "sniffio",
    "SQLAlchemy",
    "starlette",
    "streamlit",
    "sympy",
    "tavily-python==0.7.12",
    "tenacity",
    "texttable",
    "threadpoolctl",
    "tiktoken",
    "tokenizers",
    "toml",
    "torch",
    "torchvision",
    "tornado",
    "tqdm",
    "transformers",
    "triton",
    "types-PyYAML",
    "types-requests",
    "typing-inspect",
    "typing-inspection",
    "typing_extensions",
    "tzdata",
    "urllib3",
    "uvicorn",
    "uvloop",
    "virtualenv",
    "vllm",
    "wandb",
    "watchdog",
    "watchfiles",
    "websockets",
    "wheel",
    "wrapt",
    "xformers",
    "xgrammar",
    "xxhash",
    "yarl",
    "zipp",
    "zstandard",
]
